feat: #914 rename plan_viewer module to belief_viewer
