+++ src/a22_belief_viewer/static/app.js
