+++ src/a22_belief_viewer/static/style.css
